+++ blocks/card-list-block/features/card-list/default.jsx
@@ -8,8 +8,6 @@ import ArticleDate from '@wpmedia/date-block';
 import Byline from '@wpmedia/byline-block';
 import { Image } from '@wpmedia/engine-theme-sdk';
 import { extractResizedParams } from '@wpmedia/resizer-image-block';
-import PlaceholderImage from '@wpmedia/placeholder-image-block';
-
 import getProperties from 'fusion:properties';
 import './card-list.scss';
 
@@ -50,8 +48,36 @@ class CardList extends React.Component {
   constructor(props) {
     super(props);
     this.arcSite = props.arcSite;
-    this.state = { cardList: {} };
+    this.state = {
+      cardList: {},
+      placeholderResizedImageOptions: {},
+    };
     this.fetchStories();
+    this.fetchPlaceholder();
+  }
+
+  getFallbackImageURL() {
+    const { arcSite, deployment, contextPath } = this.props;
+    let targetFallbackImage = getProperties(arcSite).fallbackImage;
+
+    if (!targetFallbackImage.includes('http')) {
+      targetFallbackImage = deployment(`${contextPath}/${targetFallbackImage}`);
+    }
+
+    return targetFallbackImage;
+  }
+
+  fetchPlaceholder() {
+    const targetFallbackImage = this.getFallbackImageURL();
+
+    if (!targetFallbackImage.includes('/resources/')) {
+      this.fetchContent({
+        placeholderResizedImageOptions: {
+          source: 'resize-image-api',
+          query: { raw_image_url: targetFallbackImage, respect_aspect_ratio: true },
+        },
+      });
+    }
   }
 
   fetchStories() {
@@ -67,13 +93,17 @@ class CardList extends React.Component {
 
   render() {
     const { customFields: { title } = {}, arcSite } = this.props;
-    const { cardList: { content_elements: contentElements = [] } = {} } = this.state;
+    const {
+      cardList: { content_elements: contentElements = [] } = {},
+      placeholderResizedImageOptions,
+    } = this.state;
     const showSeparator = !!(
       contentElements[0]
       && contentElements[0].credits
       && contentElements[0].credits.by
       && contentElements[0].credits.by.length !== 0
     );
+    const targetFallbackImage = this.getFallbackImageURL();
     return (
       (contentElements.length > 0
         && (
@@ -117,13 +147,19 @@ class CardList extends React.Component {
                        resizerURL={getProperties(arcSite)?.resizerURL}
                      />
                    ) : (
-                     <PlaceholderImage
+                     <Image
                        smallWidth={377}
                        smallHeight={283}
                        mediumWidth={377}
                        mediumHeight={283}
                        largeWidth={377}
                        largeHeight={283}
+                       alt={getProperties(arcSite).primaryLogoAlt || 'Placeholder logo'}
+                       url={targetFallbackImage}
+                       breakpoints={getProperties(arcSite)?.breakpoints}
+                       resizedImageOptions={placeholderResizedImageOptions}
+                       resizerURL={getProperties(arcSite)?.resizerURL}
+
                      />
                    )
                   }
@@ -201,13 +237,18 @@ class CardList extends React.Component {
                               />
                             )
                             : (
-                              <PlaceholderImage
+                              <Image
                                 smallWidth={105}
                                 smallHeight={70}
                                 mediumWidth={105}
                                 mediumHeight={70}
                                 largeWidth={274}
                                 largeHeight={183}
+                                alt={getProperties(arcSite).primaryLogoAlt || 'Placeholder logo'}
+                                url={targetFallbackImage}
+                                breakpoints={getProperties(arcSite)?.breakpoints}
+                                resizedImageOptions={placeholderResizedImageOptions}
+                                resizerURL={getProperties(arcSite)?.resizerURL}
                               />
                             )
                         }