+++ blocks/gallery-block/features/gallery/_children/global-content.test.jsx
@@ -32,7 +32,10 @@ describe('the global content gallery', () => {
       const { default: GlobalContentGallery } = require('./global-content');
 
       const wrapper = shallow(
-        <GlobalContentGallery phrases={{ t: jest.fn((phrase) => mockPhrases[phrase]) }} />,
+        <GlobalContentGallery
+          phrases={{ t: jest.fn((phrase) => mockPhrases[phrase]) }}
+          deployment={jest.fn((path) => path)}
+        />,
       );
       expect(wrapper.find('Gallery').props().galleryElements).toStrictEqual(
         [
@@ -64,7 +67,10 @@ describe('the global content gallery', () => {
       const { default: GlobalContentGallery } = require('./global-content');
 
       const wrapper = shallow(
-        <GlobalContentGallery phrases={{ t: jest.fn((phrase) => mockPhrases[phrase]) }} />,
+        <GlobalContentGallery
+          phrases={{ t: jest.fn((phrase) => mockPhrases[phrase]) }}
+          deployment={jest.fn((path) => path)}
+        />,
       );
       expect(wrapper.find('Gallery').props().galleryElements).toStrictEqual([]);
       expect(wrapper.find('Gallery').props().expandPhrase).toEqual('Expand');
@@ -87,7 +93,10 @@ describe('the global content gallery', () => {
       const { default: GlobalContentGallery } = require('./global-content');
 
       const wrapper = shallow(
-        <GlobalContentGallery phrases={{ t: jest.fn((phrase) => mockPhrases[phrase]) }} />,
+        <GlobalContentGallery
+          phrases={{ t: jest.fn((phrase) => mockPhrases[phrase]) }}
+          deployment={jest.fn((path) => path)}
+        />,
       );
       expect(wrapper.find('Gallery').props().galleryElements).toStrictEqual([]);
       expect(wrapper.find('Gallery').props().expandPhrase).toEqual('Expand');