+++ blocks/header-nav-chain-block/chains/header-nav-chain-block/default.jsx
@@ -23,18 +23,23 @@ const sectionZIdx = navZIdx - 1;
 
 /* Styled Components */
 const StyledNav = styled.nav`
-  background-color: ${(props) => (props.navBarColor === 'light' ? '#fff' : '#000')};
-  height: ${navHeight};
-  z-index: ${navZIdx};
+  align-items: center;
+  position: relative;
+
+  .news-theme-navigation-bar {
+    background-color: ${(props) => (props.navBarColor === 'light' ? '#fff' : '#000')};
+    height: ${navHeight};
+    z-index: ${navZIdx};
+  }
 
   * {
     font-family: ${(props) => props.font};
   }
 `;
+
 const StyledSectionDrawer = styled.div`
   font-family: ${(props) => props.font};
-  position: fixed;
-  top: ${navHeight};
+  margin-top: ${navHeight};
   z-index: ${sectionZIdx};
 `;
 
@@ -74,6 +79,7 @@ const Nav = (props) => {
   const handleEscKey = (event) => {
     if (event.keyCode === 27) {
       setSectionDrawerOpen(false);
+      document.body.classList.remove('nav-open');
     }
   };
 
@@ -87,7 +93,7 @@ const Nav = (props) => {
     return () => {
       window.removeEventListener('keydown', handleEscKey);
     };
-  });
+  }, []);
 
   // Check if URL is absolute/base64
   if (primaryLogo && (primaryLogo.indexOf('http') === 0 || primaryLogo.indexOf('base64') === 0)) {
@@ -98,34 +104,36 @@ const Nav = (props) => {
 
   return (
     <>
-      <StyledNav className={`news-theme-navigation-container ${navColor === 'light' ? 'light' : 'dark'}`} font={primaryFont} navBarColor={navColor}>
+      <StyledNav id="main-nav" className={`${navColor === 'light' ? 'light' : 'dark'}`} font={primaryFont} navBarColor={navColor}>
+        <div className="news-theme-navigation-container news-theme-navigation-bar">
+          <div className="nav-left">
+            <SearchBox iconSize={20} navBarColor={navColor} placeholderText={phrases.t('header-nav-block.search-text')} />
+            <button onClick={hamburgerClick} className={`nav-btn nav-sections-btn border transparent ${navColor === 'light' ? 'nav-btn-light' : 'nav-btn-dark'}`} type="button">
+              <span>{phrases.t('header-nav-block.sections-button')}</span>
+              <HamburgerMenuIcon fill={null} height={iconSize} width={iconSize} />
+            </button>
+          </div>
 
-        <div className="nav-left">
-          <SearchBox iconSize={20} navBarColor={navColor} placeholderText={phrases.t('header-nav-chain-block.search-text')} />
-          <button onClick={hamburgerClick} className={`nav-btn nav-sections-btn border transparent ${navColor === 'light' ? 'nav-btn-light' : 'nav-btn-dark'}`} type="button">
-            <span>{phrases.t('header-nav-chain-block.sections-button')}</span>
-            <HamburgerMenuIcon fill={null} height={iconSize} width={iconSize} />
-          </button>
+          <div className="nav-logo">
+            <a href="/" title={primaryLogoAlt}>
+              {!!primaryLogo && <img src={primaryLogoPath} alt={primaryLogoAlt || 'Navigation bar logo'} />}
+            </a>
+          </div>
+
+          <div className="nav-right">
+            {signInButton}
+          </div>
         </div>
 
-        <div className="nav-logo">
-          <a href="/" title={primaryLogoAlt}>
-            {!!primaryLogo && <img src={primaryLogoPath} alt={primaryLogoAlt || 'Navigation bar logo'} />}
-          </a>
-        </div>
+        <StyledSectionDrawer id="nav-sections" className={isSectionDrawerOpen ? 'open' : 'closed'} font={primaryFont}>
+          <div className="innerDrawerNav">
+            <SectionNav sections={sections}>
+              <SearchBox alwaysOpen placeholderText={phrases.t('header-nav-block.search-text')} />
+            </SectionNav>
+          </div>
+        </StyledSectionDrawer>
 
-        <div className="nav-right">
-          {signInButton}
-        </div>
       </StyledNav>
-
-      <StyledSectionDrawer id="nav-sections" className={isSectionDrawerOpen ? 'open' : 'closed'} font={primaryFont}>
-        <SectionNav sections={sections}>
-          <SearchBox alwaysOpen placeholderText={phrases.t('header-nav-chain-block.search-text')} />
-        </SectionNav>
-      </StyledSectionDrawer>
-
-      {isSectionDrawerOpen ? <div id="overlay" role="dialog" onKeyDown={handleEscKey} onClick={() => setSectionDrawerOpen(false)} /> : null}
     </>
   );
 };