+++ blocks/numbered-list-block/features/numbered-list/default.jsx
@@ -41,12 +41,10 @@ class NumberedList extends Component {
 
 
   getFallbackImageURL() {
-    const { deployment, contextPath, arcSite } = this.props;
+    const { arcSite, deployment, contextPath } = this.props;
     let targetFallbackImage = getProperties(arcSite).fallbackImage;
 
-    // if true then it's a local image
-    // else it's a url image that can be served
-    if (targetFallbackImage && !(targetFallbackImage.includes('http'))) {
+    if (!targetFallbackImage.includes('http')) {
       targetFallbackImage = deployment(`${contextPath}/${targetFallbackImage}`);
     }
 
@@ -67,12 +65,14 @@ class NumberedList extends Component {
   fetchPlaceholder() {
     const targetFallbackImage = this.getFallbackImageURL();
 
-    this.fetchContent({
-      placeholderResizedImageOptions: {
-        source: 'resize-image-api',
-        query: { raw_image_url: targetFallbackImage, respect_aspect_ratio: true },
-      },
-    });
+    if (!targetFallbackImage.includes('/resources/')) {
+      this.fetchContent({
+        placeholderResizedImageOptions: {
+          source: 'resize-image-api',
+          query: { raw_image_url: targetFallbackImage, respect_aspect_ratio: true },
+        },
+      });
+    }
   }
 
   render() {