+++ blocks/search-results-list-block/features/search-results-list/_children/global-content.test.jsx
@@ -10,7 +10,9 @@ jest.mock('fusion:themes', () => ({
   __esModule: true,
   default: jest.fn(() => ({ 'primary-font-family': 'Open Sans', 'primary-color': '#10c8cd' })),
 }));
-jest.mock('fusion:properties', () => jest.fn(() => ({})));
+jest.mock('fusion:properties', () => (jest.fn(() => ({
+  fallbackImage: 'placeholder.jpg',
+}))));
 jest.mock('fusion:intl', () => ({
   __esModule: true,
   default: jest.fn((locale) => ({ t: jest.fn((phrase) => require('../../../intl.json')[phrase][locale]) })),
@@ -34,7 +36,8 @@ jest.mock('@wpmedia/engine-theme-sdk', () => ({
 describe('The search results list', () => {
   describe('renders a search bar', () => {
     const { default: SearchResultsList } = require('./global-content');
-    const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" />);
+    SearchResultsList.prototype.fetchContent = jest.fn();
+    const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
 
     it('should render a text input', () => {
       expect(wrapper.find('.search-bar').length).toEqual(1);
@@ -93,7 +96,8 @@ describe('The search results list', () => {
 
   it('should render a list of stories', () => {
     const { default: SearchResultsList } = require('./global-content');
-    const wrapper = shallow(<SearchResultsList globalContent={mockData} arcSite="the-sun" />);
+    SearchResultsList.prototype.fetchContent = jest.fn();
+    const wrapper = shallow(<SearchResultsList globalContent={mockData} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
     expect(wrapper.find('.results-list-container').length).toEqual(1);
     expect(wrapper.find('.list-item').length).toEqual(28);
     expect(wrapper.find('.results-list-container').childAt(0).hasClass('list-item')).toEqual(true);
@@ -101,7 +105,8 @@ describe('The search results list', () => {
 
   describe('renders one list item correctly', () => {
     const { default: SearchResultsList } = require('./global-content');
-    const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" />);
+    SearchResultsList.prototype.fetchContent = jest.fn();
+    const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
     it('should have one parent wrapper', () => {
       expect(wrapper.find('.results-list-container').length).toEqual(1);
     });
@@ -154,7 +159,8 @@ describe('The search results list', () => {
 
   describe('renders one list item correctly when description is missing', () => {
     const { default: SearchResultsList } = require('./global-content');
-    const wrapper = shallow(<SearchResultsList globalContent={LineItemWithOutDescription} arcSite="the-sun" />);
+    SearchResultsList.prototype.fetchContent = jest.fn();
+    const wrapper = shallow(<SearchResultsList globalContent={LineItemWithOutDescription} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
     it('should render one parent wrapper', () => {
       expect(wrapper.find('.results-list-container').length).toEqual(1);
     });
@@ -178,7 +184,8 @@ describe('The search results list', () => {
 
   describe('renders one list item correctly when list of authors is missing', () => {
     const { default: SearchResultsList } = require('./global-content');
-    const wrapper = shallow(<SearchResultsList globalContent={withoutByline} arcSite="the-sun" />);
+    SearchResultsList.prototype.fetchContent = jest.fn();
+    const wrapper = shallow(<SearchResultsList globalContent={withoutByline} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
     it('should render one parent wrapper', () => {
       expect(wrapper.find('.results-list-container').length).toEqual(1);
     });
@@ -191,19 +198,22 @@ describe('The search results list', () => {
   describe('renders a button to display more stories', () => {
     it('should render a button to display more stories', () => {
       const { default: SearchResultsList } = require('./global-content');
-      const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" />);
+      SearchResultsList.prototype.fetchContent = jest.fn();
+      const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
       expect(wrapper.find('.see-more').childAt(0).length).toEqual(1);
     });
 
     it('should have the primary text as font family', () => {
       const { default: SearchResultsList } = require('./global-content');
-      const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" />);
+      SearchResultsList.prototype.fetchContent = jest.fn();
+      const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
       expect((wrapper.find('.see-more')).childAt(0)).toHaveProp('primaryFont', 'Open Sans');
     });
 
     it('should have the primary text as font family', () => {
       const { default: SearchResultsList } = require('./global-content');
-      const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" />);
+      SearchResultsList.prototype.fetchContent = jest.fn();
+      const wrapper = shallow(<SearchResultsList globalContent={oneListItem} arcSite="the-sun" deployment={jest.fn((path) => path)} />);
       expect((wrapper.find('.see-more')).childAt(0)).toHaveProp('primaryColor', '#10c8cd');
     });