+++ blocks/simple-list-block/features/simple-list/default.jsx
@@ -1,11 +1,12 @@
 /* eslint-disable camelcase */
-import React from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { useContent } from 'fusion:content';
 import { useFusionContext } from 'fusion:context';
 import { extractResizedParams } from '@wpmedia/resizer-image-block';
 import getThemeStyle from 'fusion:themes';
 import getProperties from 'fusion:properties';
+import Consumer from 'fusion:consumer';
 import Title from './_children/title';
 import StoryItem from './_children/story-item';
 import './simple-list.scss';
@@ -26,6 +27,56 @@ const unserializeStory = (storyObject) => ({
 
 // helpers end
 
+@Consumer
+class SimpleListWrapper extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      placeholderResizedImageOptions: {},
+    };
+    this.fetchPlaceholder();
+  }
+
+  getFallbackImageURL() {
+    const { arcSite, deployment, contextPath } = this.props;
+    let targetFallbackImage = getProperties(arcSite).fallbackImage;
+
+    if (!targetFallbackImage.includes('http')) {
+      targetFallbackImage = deployment(`${contextPath}/${targetFallbackImage}`);
+    }
+
+    return targetFallbackImage;
+  }
+
+  fetchPlaceholder() {
+    const targetFallbackImage = this.getFallbackImageURL();
+
+    // using the fetchContent seems both more reliable
+    // and allows for conditional calls whereas useContent hook does not
+    if (!targetFallbackImage.includes('/resources/')) {
+      this.fetchContent({
+        placeholderResizedImageOptions: {
+          source: 'resize-image-api',
+          query: { raw_image_url: targetFallbackImage, respect_aspect_ratio: true },
+        },
+      });
+    }
+  }
+
+  render() {
+    const { placeholderResizedImageOptions } = this.state;
+    const targetFallbackImage = this.getFallbackImageURL();
+    return (
+      <SimpleList
+        // eslint-disable-next-line react/jsx-props-no-spreading
+        {...this.props}
+        placeholderResizedImageOptions={placeholderResizedImageOptions}
+        targetFallbackImage={targetFallbackImage}
+      />
+    );
+  }
+}
+
 const SimpleList = (props) => {
   const {
     customFields: {
@@ -38,6 +89,8 @@ const SimpleList = (props) => {
       showImage = true,
     } = {},
     id = '',
+    placeholderResizedImageOptions,
+    targetFallbackImage,
   } = props;
 
   const { arcSite } = useFusionContext();
@@ -73,6 +126,8 @@ const SimpleList = (props) => {
             showHeadline={showHeadline}
             showImage={showImage}
             resizedImageOptions={resizedImageOptions}
+            placeholderResizedImageOptions={placeholderResizedImageOptions}
+            targetFallbackImage={targetFallbackImage}
           />
         ))
       }
@@ -80,7 +135,7 @@ const SimpleList = (props) => {
   );
 };
 
-SimpleList.propTypes = {
+SimpleListWrapper.propTypes = {
   customFields: PropTypes.shape({
     listContentConfig: PropTypes.contentConfig('ans-feed').tag(
       {
@@ -102,6 +157,6 @@ SimpleList.propTypes = {
   }),
 };
 
-SimpleList.label = 'Simple List – Arc Block';
+SimpleListWrapper.label = 'Simple List – Arc Block';
 
-export default SimpleList;
+export default SimpleListWrapper;