+++ blocks/top-table-list-block/features/top-table-list/_children/horizontal-overline-image-story-item.test.jsx
@@ -3,7 +3,12 @@ import { mount } from 'enzyme';
 
 jest.mock('fusion:properties', () => (jest.fn(() => ({
   fallbackImage: 'placeholder.jpg',
+  resizerURL: 'resizer',
 }))));
+jest.mock('@wpmedia/engine-theme-sdk', () => ({
+  Image: () => <img alt="test" />,
+}));
+
 const config = {
   showOverlineXL: true,
   showHeadlineXL: true,