+++ blocks/footer-block/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.18](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/footer-block@2.0.17...@wpmedia/footer-block@2.0.18) (2020-06-02)
+
+
+### Reverts
+
+* Revert "Remove references to resized images" ([c9e6993](https://github.com/WPMedia/fusion-news-theme-blocks/commit/c9e6993c4c3facdac967fa3117ff71309beaae8d))
+
+
+
+
+
 ## [2.0.17](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/footer-block@2.0.17-hotfix.0...@wpmedia/footer-block@2.0.17) (2020-05-15)
 
 **Note:** Version bump only for package @wpmedia/footer-block