+++ blocks/header-nav-block/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.3](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/header-nav-block@2.3.3-beta.1...@wpmedia/header-nav-block@2.3.3) (2020-06-02)
+
+**Note:** Version bump only for package @wpmedia/header-nav-block
+
+
+
+
+
 ## [2.2.1](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/header-nav-block@2.2.1-hotfix.0...@wpmedia/header-nav-block@2.2.1) (2020-05-15)
 
 **Note:** Version bump only for package @wpmedia/header-nav-block