+++ blocks/site-hierarchy-content-block/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.14](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/site-hierarchy-content-block@3.1.13...@wpmedia/site-hierarchy-content-block@3.1.14) (2020-06-02)
+
+
+### Reverts
+
+* Revert "Remove references to resized images" ([c9e6993](https://github.com/WPMedia/fusion-news-theme-blocks/commit/c9e6993c4c3facdac967fa3117ff71309beaae8d))
+
+
+
+
+
 ## [3.1.13](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/site-hierarchy-content-block@3.1.13-hotfix.0...@wpmedia/site-hierarchy-content-block@3.1.13) (2020-05-15)
 
 **Note:** Version bump only for package @wpmedia/site-hierarchy-content-block