+++ blocks/small-promo-block/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/small-promo-block@1.2.4-beta.0...@wpmedia/small-promo-block@1.2.4) (2020-06-02)
+
+**Note:** Version bump only for package @wpmedia/small-promo-block
+
+
+
+
+
 ## [1.2.3](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/small-promo-block@1.2.3-hotfix.0...@wpmedia/small-promo-block@1.2.3) (2020-05-15)
 
 **Note:** Version bump only for package @wpmedia/small-promo-block